Move min/max prefix length validators to ipam.validators
